Move AnkiProgressBar class into its own script
